+++ spec/fixtures/boot_apps/syntax_error/config/application.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module BrokenSyntax
+  class Application < Rails::Application
+end